+++ client/src/components/pages/Kurate/Blog/Blog.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Posts from '../Posts'
+
+const Blog = (props) => (
+  <Posts
+    match={props.match}
+  />
+)
+
+export default Blog;
