+++ client/src/components/pages/Kurate/Feed/Feed.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Posts from '../Posts'
+
+const Feed = (props) => (
+  <Posts
+    match={props.match}
+  />
+)
+
+export default Feed;
